feat: 完善 API 请求

+++ src/comments/comments.controller.ts
@@ -7,7 +7,9 @@ import {
   Patch,
   Post,
   Query,
+  UnauthorizedException,
 } from '@nestjs/common';
+import { Throttle } from '@nestjs/throttler';
 
 import { CommentsService } from './comments.service';
 import { CreateCommentDto } from './dto/create-comment.dto';
@@ -22,12 +24,9 @@ export class CommentsController {
     return this.commentsService.create(data);
   }
 
+  @Throttle({ default: { limit: 100, ttl: 60000 } })
   @Get()
-  async findAll(@Query('path') path: string, @Query('admin') admin: number) {
-    if (admin) {
-      return await this.commentsService.adminUpdate();
-    }
-
+  async findAll(@Query('path') path: string) {
     if (path) {
       return await this.commentsService.findAllByPath(path);
     } else {
@@ -35,18 +34,31 @@ export class CommentsController {
     }
   }
 
+  @Throttle({ default: { limit: 100, ttl: 60000 } })
   @Get(':id')
   async findOne(@Param('id') id: number) {
     return this.commentsService.findOne(id);
   }
 
   @Patch(':id')
-  async update(@Param('id') id: number, @Body() data: UpdateCommentDto) {
+  async update(
+    @Param('id') id: number,
+    @Body() data: UpdateCommentDto,
+    @Query('token') token: string,
+  ) {
+    if (token !== process.env.SECRET_KEY) {
+      throw new UnauthorizedException('非法请求');
+    }
+
     return this.commentsService.update(id, data);
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: number) {
+  async remove(@Param('id') id: number, @Query('token') token: string) {
+    if (token !== process.env.SECRET_KEY) {
+      throw new UnauthorizedException('非法请求');
+    }
+
     return this.commentsService.remove(id);
   }
 }
